backend/pyproject.toml
[project]
name = "isro-vision-api"
version = "0.1.0"
description = "ISRO Vision API - Multimodal Chatbot Orchestrator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.1",
    "passlib>=1.7.4",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=1.10.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.5",
    "motor>=3.0.0",
    "requests>=2.31.0",
    "pymongo>=4.0.0",
    "email-validator>=2.0.0",
    "boto3>=1.28.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

